ex_getln: save the commandline in getexline too
It's possible to recurse in getexline, which also results in recursing
into getcmdline.

getcmdline can't be safely recursed into without manually updating
ccline.

This can specifically occur when an event on the queue triggers it (such
as if a plugin registers a job callback, and then does an 'exec' within
its callback).

This adds the required boilerplate to getexline such that it's safe to
recurse into getcmdline.

It's possible that the save/restore could be higher up the stack and
thus impact fewer total callers, but I decided to err on the side of
correctness.

Fixes a possible segfault, see #6023